requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brotli==1.0.9
cffi==1.15.0
click==8.1.2
dnspython==2.2.1
email-validator==1.1.3
Faker==13.3.4
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Compress==1.11
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Limiter==2.2.0
Flask-Mail==0.9.1
Flask-Mail-SendGrid==0.3
Flask-PyMongo==2.3.0
Flask-Share==0.1.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
limits==2.4.0
MarkupSafe==2.1.1
mongoengine==0.24.1
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.11.4
python-dateutil==2.8.2
python-http-client==3.3.7
python-slugify==6.1.1
pytz==2022.1
sendgrid==6.9.7
six==1.16.0
soupsieve==2.3.2.post1
starkbank-ecdsa==2.0.3
text-unidecode==1.3
validate-email==1.3
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
